Refactor translators to remove super.translate
The previous type hierarchy required that OMETranslators called
super.typedTranslate in their own typedTranslate implementation. This
was very error prone, and indeed several translators did not call
super.typedTranslate and were thus broken.

To move away from this reliance on convention, the OMETranslator class
now introduces a new moethod - translateOMEXML. In the ToOMETranslator
and FromOMETranslators, the typedTranslate method is now implemented and
delegates to translateOMEXML. Thus, concrete implementations have to
implement translateOMEXML instead of typedTranslate now.

The logic to populate OMEXML was migrated from the DefaultOMETranslator
to the ToOMEXMLTranslator class, as it was only relevant when populating
OMEMetadata objects (not when reading from them). This also removed the
necessity to chain translate, which should be safer in general.

Closes #4