OMETIFFFormat: remove static parameters
The parent OMETIFFFormat class was declaring static @Parameters, which
were then used by all component classes. This is terrible design, as if
multiple contexts are created, subsequent OMETIFFFormats will throw
IllegalStateExceptions when the context is attempted to be injected
(because the fields are static).

There is no reason not to use the framework here and move the
@Parameters to each class that actually uses them.

Closes #19