Remove translateOMEXML signature
This method, if not unnecessary from the beginning, was certainly made
so by splitting Metadata and ImageMetadata translation signatures. The
Translator is now easier to extend than when these classes were
originally written.

For the bulk of these translators it is sufficient to put the
translation in the Format metadata translation method.

ImageMetadata population can now be handled in its appropriate
method, using the new OMEXMLMetadataService population signature.